refactor: Receiver 컴포넌트 기능별 분리
